+++ RentACarSpring/src/main/java/com/etiya/rentACarSpring/businnes/concretes/CarDamageManager.java
@@ -11,11 +11,11 @@ import com.etiya.rentACarSpring.core.utilities.mapping.ModelMapperService;
 import com.etiya.rentACarSpring.core.utilities.results.*;
 import com.etiya.rentACarSpring.dataAccess.abstracts.CarDamageDao;
 
-import com.etiya.rentACarSpring.entities.Car;
 import com.etiya.rentACarSpring.entities.CarDamage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -57,7 +57,8 @@ public class CarDamageManager implements CarDamageService {
 
     @Override
     public Result update(UpdateCarDamageRequest updateCarDamageRequest) {
-        Result result = BusinnessRules.run(checkCarExistsInGallery(updateCarDamageRequest.getCarId()));
+        Result result = BusinnessRules.run(checkCarExistsInGallery(updateCarDamageRequest.getCarId()),
+                checkIfCarDamageIdExists(updateCarDamageRequest.getCarDamageId()));
         if (result != null) {
             return result;
         }
@@ -68,14 +69,28 @@ public class CarDamageManager implements CarDamageService {
     }
 
     @Override
-    public Result delete(DeleteCarDamageRequest deleteCarDamageRequest) {
+    public Result delete(DeleteCarDamageRequest deleteCarDamageRequest) throws EntityNotFoundException {
+        Result result = BusinnessRules.run(
+                checkIfCarDamageIdExists(deleteCarDamageRequest.getCarDamageId()));
+        if (result != null) {
+            return result;
+        }
+
         this.carDamageDao.deleteById(deleteCarDamageRequest.getCarDamageId());
         return new SuccesResult("Silindi");
     }
 
+    @Override
+    public Result checkIfCarDamageIdExists(int carDamageId) {
+        if (!this.carDamageDao.existsById(carDamageId)) {
+            return new ErrorResult("carDamageId Id bulunamadı");
+        }
+        return new SuccesResult();
+    }
+
     private Result checkCarExistsInGallery(int id) {
         boolean isExisting = carService.checkCarExistsInGallery(id).isSuccess();
-        if (!isExisting) {
+        if (isExisting) {
             return new SuccesResult();
         }
         return new ErrorResult("Galeride böyle bir araba yok.");
